Move TreeGrid specific demos to its own class (#609)
That way the TreeGrid demos can be isolated in vaadin.com.
