+++ vaadin-grid-flow-demo/src/main/java/com/vaadin/flow/component/treegrid/demo/HierarchicalTestBean.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License. 
  */
-package com.vaadin.flow.component.grid.demo;
+package com.vaadin.flow.component.treegrid.demo;
 
 public class HierarchicalTestBean {
